Carrying the torch of CoCoWiWa into the realm of VS
Mostly signed vs unsigned and double to float narrowing fixes.
Adding Material class for physics.

+++ src/engine/physics/defines.hpp
@@ -7,8 +7,8 @@ namespace engine
 namespace physics
 {
 	typedef std::size_t Id;
-	typedef std::array<double, 3> Point;
-	typedef std::array<double, 3> Size;
-	typedef std::array<double, 2> Vec2;
+	typedef std::array<float, 3> Point;
+	typedef std::array<float, 3> Size;
+	typedef std::array<float, 2> Vec2;
 }
 }
